requirements.txt
requests
rich
jinja2
pyyaml
openapi-spec-validator
flask
